Moved code in floors/floors.go
