+++ config/config.go
@@ -1209,9 +1209,9 @@ func SetupViper(v *viper.Viper, filename string) {
 	v.SetDefault("account_defaults.price_floors.enabled", true)
 	v.SetDefault("account_defaults.price_floors.enforce_floors_rate", 100)
 	v.SetDefault("account_defaults.price_floors.adjust_for_bid_adjustment", true)
-	v.SetDefault("account_defaults.price_floors.enforce_deal_floors", false)
+	v.SetDefault("account_defaults.price_floors.enforce_deal_floors", true)
 	v.SetDefault("account_defaults.price_floors.use_dynamic_data", true)
-	v.SetDefault("account_defaults.price_floors.fetch.enabled", false)
+	v.SetDefault("account_defaults.price_floors.fetch.enabled", true)
 	v.SetDefault("account_defaults.price_floors.fetch.timeout_ms", 3000)
 	v.SetDefault("account_defaults.price_floors.fetch.max_file_size_kb", 100)
 	v.SetDefault("account_defaults.price_floors.fetch.max_rules", 1000)
